refactor: break up config list cells to own files and refactor
refactor: break the config list cells component to own files

refactor: break the config list cells component to own files

fix: fix type issue on configs tags dropdown

refactor: remove unnecessary console logs

refactor: move config list to file on its own
